+++ NhakhoaMyNgoc_Connector/DTOs/SummaryServiceDto.cs
@@ -10,8 +10,8 @@ namespace NhakhoaMyNgoc_Connector.DTOs
 
         public string ServiceName { get; set; }
         public int Quantity { get; set; }
-        public int Price { get; set; }
-        public int Discount { get; set; }
-        public int Total { get; set; }
+        public long Price { get; set; }
+        public long Discount { get; set; }
+        public long Total { get; set; }
     }
 }
